Ensuring every library depends on `strict`.
Some strict issue have crept in and went unnoticed.

I observed that the generated dependency graph was not rooted at `strict`. Fixes have been included to make sure that strict rules are applied to all code.

+++ modules/stm32/source/Pin.cpp
@@ -14,9 +14,9 @@ Pin::Pin(Port port, uint8_t index)
 }
 
 Pin& Pin::SetMode(Mode mode) {
-    auto pin_mode = general_purpose_input_output[to_underlying(port_)].mode;    // read
-    pin_mode.whole |= (to_underlying(mode) << (index_ * 2U));                   // modify
-    general_purpose_input_output[to_underlying(port_)].mode = pin_mode;         // write
+    auto pin_mode = general_purpose_input_output[to_underlying(port_)].mode;               // read
+    pin_mode.whole |= static_cast<std::uint32_t>(to_underlying(mode) << (index_ * 2U));    // modify
+    general_purpose_input_output[to_underlying(port_)].mode = pin_mode;                    // write
     return *this;
 }
 
@@ -26,9 +26,9 @@ Mode Pin::GetMode() const {
 }
 
 Pin& Pin::SetOutputSpeed(Speed speed) {
-    auto output_speed = general_purpose_input_output[to_underlying(port_)].output_speed;    // read
-    output_speed.whole |= (to_underlying(speed) << (index_ * 2U));                          // modify
-    general_purpose_input_output[to_underlying(port_)].output_speed = output_speed;         // write
+    auto output_speed = general_purpose_input_output[to_underlying(port_)].output_speed;        // read
+    output_speed.whole |= static_cast<std::uint32_t>(to_underlying(speed) << (index_ * 2U));    // modify
+    general_purpose_input_output[to_underlying(port_)].output_speed = output_speed;             // write
     return *this;
 }
 
@@ -39,7 +39,7 @@ Speed Pin::GetOutputSpeed() const {
 
 Pin& Pin::SetOutputType(OutputType type) {
     auto output_type = general_purpose_input_output[to_underlying(port_)].output_type;    // read
-    output_type.whole |= (to_underlying(type) << index_);                                 // modify
+    output_type.whole |= static_cast<std::uint32_t>(to_underlying(type) << index_);       // modify
     general_purpose_input_output[to_underlying(port_)].output_type = output_type;         // write
     return *this;
 }
@@ -51,7 +51,7 @@ OutputType Pin::GetOutputType() const {
 
 Pin& Pin::SetResistor(Resistor type) {
     auto pull = general_purpose_input_output[to_underlying(port_)].pullup_pulldown;    // read
-    pull.whole |= (to_underlying(type) << (index_ * 2U));                              // modify
+    pull.whole |= static_cast<std::uint32_t>(to_underlying(type) << (index_ * 2U));    // modify
     general_purpose_input_output[to_underlying(port_)].pullup_pulldown = pull;         // write
     return *this;
 }
@@ -77,7 +77,7 @@ void Pin::Value(bool value) {
     Mode mode = GetMode();
     if (mode == Mode::Output) {
         auto output = general_purpose_input_output[to_underlying(port_)].output_data;    // read
-        output.whole |= (value << index_);                                               // modify
+        output.whole |= static_cast<std::uint32_t>(value << index_);                     // modify
         general_purpose_input_output[to_underlying(port_)].output_data = output;         // write
     } else {
         // nothing happens
@@ -90,11 +90,11 @@ Pin& Pin::SetAlternative(uint8_t value) {
         value &= 0xFU;    // only 4 bits are allowed
         if (index_ < 8) {
             auto alt_func_low = general_purpose_input_output[to_underlying(port_)].alt_func_low;    // read
-            alt_func_low.whole |= (value << (index_ * 4U));                                         // modify
+            alt_func_low.whole |= static_cast<std::uint32_t>(value << (index_ * 4U));               // modify
             general_purpose_input_output[to_underlying(port_)].alt_func_low = alt_func_low;         // write
         } else {
             auto alt_func_high = general_purpose_input_output[to_underlying(port_)].alt_func_high;    // read
-            alt_func_high.whole |= (value << ((index_ - 8) * 4U));                                    // modify
+            alt_func_high.whole |= static_cast<std::uint32_t>(value << ((index_ - 8) * 4U));          // modify
             general_purpose_input_output[to_underlying(port_)].alt_func_high = alt_func_high;         // write
         }
     }
